set up private route for search, redirecting to search after sign up or sign in, and other minor changes
