+++ src/components/SignUp.js
@@ -41,7 +41,7 @@ const SignUp = props => {
 			.then(response => {
 				console.log(response);
 				localStorage.setItem('user_id', response.data.user.id);
-				props.history.push('/dashboard');
+				props.history.push('/search');
 			})
 			.catch(error => console.log(error));
 	};
