+++ src/utils/PrivateRoute.js
@@ -9,7 +9,7 @@ export const PrivateRoute = ({ component: Component, ...rest }) => {
 				if (localStorage.getItem('user_id')) {
 					return <Component {...props} />;
 				} else {
-					return <Redirect to="/failure" />;
+					return <Redirect to="/pagenotfound" />;
 				}
 			}}
 		/>
